requirements.txt
Flask==3.0.0
joblib==1.3.2
numpy==1.24.3
scikit-learn==1.3.2
lightgbm==4.1.0
google-generativeai
